Update README.md
You have recently just been hired as Java Football Club’s new manager and are already aiming your sights to win the league cup this season. You didn’t expect that coaching a soccer team would be so difficult, so you decided to use your knowledge of object-oriented programming to write a program to help you keep track of your players. To do this, you will create four classes: Position.java, Player.java, and Roster.java to put together your team, and Driver.java to test your code.

Solution Description

Create Position.java to define the various positions on the field that a player can play, Player.java to represent the players you coach, and Roster.java to represent the roster of players on Java FC. You will be creating several fields and methods for each class. Based on the description given for each field and method, you will have to decide whether they should be static or non-static, and whether it should be private or public. To make these decisions, you should carefully follow the guidelines on these keywords as taught in lecture. In some cases, your program will still function with an incorrect keyword.

REMEMBER: Homeworks are different from PEs in that they state REQUIREMENTS. These requirements may NOT always be stated in the ORDER in which you should implement them. For example, we may ask you to return a value and report an error, but these two requirements must be handled in the correct order.

Notes:

All variables must not be allowed to be directly modified outside the class in which they are declared and require an instance to be accessed through, unless otherwise stated in the description of the variable.

All methods should be accessible from everywhere and require an instance to be accessed through, unless otherwise specified.

Use constructor and method chaining whenever possible! Ensure that your constructor chaining helps you reduce code repetition and maximize code reuse.